% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec}
\alias{cec}
\title{CEC interface}
\usage{
cec(func_index, x, cec, max_func_index, dims, suite = NULL)
}
\arguments{
\item{func_index}{numeric index of optimisation problem in given benchmark}

\item{x}{vector of numeric inputs for objective function}

\item{cec}{name of benchmark}

\item{max_func_index}{biggest index of optimisation problem in given benchmark}

\item{dims}{vector of available dimensionalities}

\item{suite}{one of the suite in CEC2021 benchmark (basic, shift, rot, bias, shift_rot, bias_rot, bias_shift, bias_shift_rot)}
}
\value{
value of objective function for given input
}
\description{
The common interface for all available benchmark from CEC.
}
