% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec2014}
\alias{cec2014}
\title{CEC2014 interface}
\usage{
cec2014(func_index, x)
}
\arguments{
\item{func_index}{numeric index of optimisation problem from set set 1:30}

\item{x}{vector of numeric inputs for objective function}
}
\value{
value of objective function for given input
}
\description{
The R interface for CEC2014 Single Objective Bound Constrained Numerical Optimization benchmark.
Available dimensions are following: (10, 20, 30, 50, 100).
}
