% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec2015}
\alias{cec2015}
\title{CEC2015 interface}
\usage{
cec2015(func_index, x)
}
\arguments{
\item{func_index}{numeric index of optimisation problem from set set 1:15}

\item{x}{vector of numeric inputs for objective function}
}
\value{
value of objective function for given input
}
\description{
The R interface for CEC2015 Single Objective Bound
Constrained Numerical Optimization benchmark.
Available dimensions are following: (10, 30, 50, 100)
}
