% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec2017}
\alias{cec2017}
\title{CEC2017 interface}
\source{
http://staff.elka.pw.edu.pl/~djagodzi/programy.html
}
\usage{
cec2017(i, x)
}
\arguments{
\item{x}{vector of numeric inputs for objective function}

\item{func_index}{numeric index of optimisation problem in given benchmark}
}
\value{
value of objective function for given input
}
\description{
The R interface for CEC2017 Single Objective Bound Constrained Numerical Optimization benchmark.
}
