% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec2019}
\alias{cec2019}
\title{CEC2019 interface}
\usage{
cec2019(func_index, x)
}
\arguments{
\item{func_index}{numeric index of optimisation problem from set set 1:10}

\item{x}{vector of numeric inputs for objective function}
}
\value{
value of objective function for given input
}
\description{
The R interface for CEC2019 100-Digit Challenge
Constrained Numerical Optimization benchmark.
Available dimensions are following: functions F1-F3 are available only for
(respective) dimensions 9, 16, and 18. Functions F4-F10 are available for
10 dimensions.
}
