% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec2021}
\alias{cec2021}
\title{CEC2021 interface}
\usage{
cec2021(func_index, x, suite)
}
\arguments{
\item{func_index}{numeric index of optimisation problem in given benchmark}

\item{x}{vector of numeric inputs for objective function}

\item{suite}{one of the suite in CEC2021 benchmark (basic, shift, rot, bias, shift_rot, bias_rot, bias_shift, bias_shift_rot)}
}
\value{
value of objective function for given input
}
\description{
The R interface for CEC2021 Single Objective Bound Constrained Numerical Optimization benchmark.
}
