% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cecs_tmp_dirpath}
\alias{cecs_tmp_dirpath}
\title{cecs temporary directory name}
\usage{
cecs_tmp_dirpath(os_name)
}
\arguments{
\item{os_name}{OS name}
}
\description{
Function returns a temporary directory name which
the package utilizes as a data storage on given
platform.
}
