% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean}
\alias{clean}
\title{Remove CEC data}
\usage{
clean()
}
\description{
Function deletes directory with CEC's data.
Some text files are quite large (even 10MB)
and if one won't use CEC's data in the near
future this function allows to free the disk space.
}
