% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{destination_file}
\alias{destination_file}
\title{Download destination}
\usage{
destination_file(os_name)
}
\arguments{
\item{os_name}{OS name}
}
\description{
Function returns filepath to the download
destination on given platform.
}
