% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Download CEC data}
\usage{
download_data(cec, os_name)
}
\arguments{
\item{cec}{name of benchmark}

\item{os_name}{OS name}
}
\description{
Function downloads numeric data for specified CEC
benchmark.
For further details, see \url{https://github.com/ewarchul/cec/tree/main/data}
}
