% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_cec_dirpath}
\alias{get_cec_dirpath}
\title{Get path to CEC data}
\usage{
get_cec_dirpath(cec)
}
\arguments{
\item{cec}{name of benchmark}
}
\description{
If temporary directory with CEC's data exists then
function returns a filepath to data of specified
CEC version. Otherwise, function creates a temporary
directory, downloads an archive with data, unzip it,
and returns mentioned datapath.
}
