% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_data}
\alias{unzip_data}
\title{Extract compressed data}
\usage{
unzip_data(archivepath)
}
\arguments{
\item{archivepath}{path to CEC's data}
}
\description{
Function extracts CEC's data from ZIP archive.
}
