% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec2013}
\alias{cec2013}
\title{CEC2013 interface}
\source{
https://github.com/hzambran/cec2013
}
\usage{
cec2013(func_index, x)
}
\arguments{
\item{func_index}{numeric index of optimisation problem in given benchmark}

\item{x}{vector of numeric inputs for objective function}
}
\value{
value of objective function for given input
}
\description{
The R interface for CEC2013 Single Objective Bound Constrained Numerical Optimization benchmark.
}
