% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{cec2017}
\alias{cec2017}
\title{CEC2017 interface}
\source{
http://staff.elka.pw.edu.pl/~djagodzi/programy.html
}
\usage{
cec2017(func_index, x)
}
\arguments{
\item{func_index}{numeric index of optimisation problem from set 1:30}

\item{x}{vector of numeric inputs for objective function}
}
\value{
value of objective function for given input
}
\description{
The R interface for CEC2017 Single Objective Bound
Constrained Numerical Optimization benchmark.
Available dimensions are following: (10, 30, 50, 100).
}
