% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cec.R
\name{clean}
\alias{clean}
\title{Remove TXT data}
\usage{
clean()
}
\description{
Function deletes decompressed directory with
CEC's data. Some text files are quite large (even 10MB)
and if one won't use specific CEC version in
the near future this function allows to free the disk space.
}
