% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{download_data}
\alias{download_data}
\title{Download CEC data}
\usage{
download_data(cec)
}
\arguments{
\item{cec}{name of benchmark}
}
\description{
Function downloads numeric data forspecified CEC
benchmark.
For further details, see \url{http://home.elka.pw.edu.pl/~ewarchul/}
}
