% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unzip_data}
\alias{unzip_data}
\title{Extract compressed data}
\usage{
unzip_data(archivepath, cec)
}
\arguments{
\item{archivepath}{path to CEC's data}

\item{cec}{name of benchmark}
}
\description{
Function extracts CEC's data from ZIP archive.
If data is already extracted function does nothing.
}
